Rename isoreg to isoreg_impl
