+++ src/scores/continuous/__init__.py
@@ -2,7 +2,7 @@
 Import the functions from the implementations into the public API
 """
 from scores.continuous.flip_flop_impl import flip_flop_index
-from scores.continuous.isoreg import isotonic_fit
+from scores.continuous.isoreg_impl import isotonic_fit
 from scores.continuous.murphy_impl import murphy_score, murphy_thetas
 from scores.continuous.quantile_loss_impl import quantile_score
 from scores.continuous.standard_impl import mae, mse, rmse
