+++ src/scores/probability/__init__.py
@@ -1,7 +1,7 @@
 """
 Import the functions from the implementations into the public API
 """
-from scores.continuous.isoreg import isotonic_fit
+from scores.continuous.isoreg_impl import isotonic_fit
 from scores.continuous.murphy_impl import murphy_score, murphy_thetas
 from scores.probability.crps_impl import (
     adjust_fcst_for_crps,
